geo_clipper

Trait Clipper

pub trait Clipper<F: CoordFloat = f64> {
    // Required methods
    fn difference<T: ToOwnedPolygon<F> + ClosedPoly + ?Sized>(
        &self,
        other: &T,
        factor: F,
    ) -> MultiPolygon<F>;
    fn intersection<T: ToOwnedPolygon<F> + ClosedPoly + ?Sized>(
        &self,
        other: &T,
        factor: F,
    ) -> MultiPolygon<F>;
    fn union<T: ToOwnedPolygon<F> + ClosedPoly + ?Sized>(
        &self,
        other: &T,
        factor: F,
    ) -> MultiPolygon<F>;
    fn xor<T: ToOwnedPolygon<F> + ClosedPoly + ?Sized>(
        &self,
        other: &T,
        factor: F,
    ) -> MultiPolygon<F>;
    fn offset(
        &self,
        delta: F,
        join_type: JoinType,
        end_type: EndType,
        factor: F,
    ) -> MultiPolygon<F>;
    fn offset_simplify_clean(
        &self,
        delta: F,
        jt: JoinType,
        et: EndType,
        pft: PolyFillType,
        distance: F,
        factor: F,
    ) -> MultiLineString<F>;
    fn simplify(&self, fill_type: PolyFillType, factor: F) -> MultiLineString<F>;
    fn clean(&self, distance: F, factor: F) -> MultiLineString<F>;
}

This trait defines the boolean and offset operations on polygons

The factor parameter in its methods is used to scale shapes before and after applying the operation to avoid precision loss since Clipper (the underlaying library) performs integer computation.

Required Methods

fn difference<T: ToOwnedPolygon<F> + ClosedPoly + ?Sized>( &self, other: &T, factor: F, ) -> MultiPolygon<F>

fn intersection<T: ToOwnedPolygon<F> + ClosedPoly + ?Sized>( &self, other: &T, factor: F, ) -> MultiPolygon<F>

fn union<T: ToOwnedPolygon<F> + ClosedPoly + ?Sized>( &self, other: &T, factor: F, ) -> MultiPolygon<F>

fn xor<T: ToOwnedPolygon<F> + ClosedPoly + ?Sized>( &self, other: &T, factor: F, ) -> MultiPolygon<F>

fn offset( &self, delta: F, join_type: JoinType, end_type: EndType, factor: F, ) -> MultiPolygon<F>

fn offset_simplify_clean( &self, delta: F, jt: JoinType, et: EndType, pft: PolyFillType, distance: F, factor: F, ) -> MultiLineString<F>

fn simplify(&self, fill_type: PolyFillType, factor: F) -> MultiLineString<F>

fn clean(&self, distance: F, factor: F) -> MultiLineString<F>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.