geo_clipper

Trait ClipperOpenInt

pub trait ClipperOpenInt {
    // Required methods
    fn difference<T: ToOwnedPolygonInt + ClosedPoly + ?Sized>(
        &self,
        other: &T,
    ) -> MultiLineString<i64>;
    fn intersection<T: ToOwnedPolygonInt + ClosedPoly + ?Sized>(
        &self,
        other: &T,
    ) -> MultiLineString<i64>;
    fn offset(
        &self,
        delta: f64,
        join_type: JoinType,
        end_type: EndType,
    ) -> MultiPolygon<i64>;
}

This trait defines the boolean and offset operations between open paths and polygons, for integer coordinate types It is a subset of the operations for polygons

There is no factor, since polygons are already in integer form

Required Methods

fn difference<T: ToOwnedPolygonInt + ClosedPoly + ?Sized>( &self, other: &T, ) -> MultiLineString<i64>

fn intersection<T: ToOwnedPolygonInt + ClosedPoly + ?Sized>( &self, other: &T, ) -> MultiLineString<i64>

fn offset( &self, delta: f64, join_type: JoinType, end_type: EndType, ) -> MultiPolygon<i64>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.